chore: Rename DigitalPlanningApplication to Application
