+++ examples/application/data/listedBuildingConsent.ts
@@ -1,8 +1,8 @@
-import {DigitalPlanningApplication} from '../../../types/schemas/digitalPlanningApplication';
+import {Application} from '../../../types/schemas/application';
 
 const version = process.env['VERSION'] || '@next';
 
-export const listedBuildingConsent: DigitalPlanningApplication = {
+export const listedBuildingConsent: Application = {
   data: {
     user: {
       role: 'applicant',
