+++ examples/application/data/planningPermission/fullHouseholder.ts
@@ -1,8 +1,8 @@
-import {DigitalPlanningApplication} from '../../../../types/schemas/digitalPlanningApplication';
+import {Application} from '../../../../types/schemas/application';
 
 const version = process.env['VERSION'] || '@next';
 
-export const planningPermissionFullHouseholder: DigitalPlanningApplication = {
+export const planningPermissionFullHouseholder: Application = {
   data: {
     application: {
       type: {
