+++ examples/application/data/priorApproval/largerExtension.ts
@@ -1,8 +1,8 @@
-import {DigitalPlanningApplication} from '../../../../types/schemas/digitalPlanningApplication';
+import {Application} from '../../../../types/schemas/application';
 
 const version = process.env['VERSION'] || '@next';
 
-export const priorApprovalLargerExtension: DigitalPlanningApplication = {
+export const priorApprovalLargerExtension: Application = {
   data: {
     user: {
       role: 'applicant',
