+++ examples/application/planningPermission/fullHouseholder.json
