+++ examples/application/planningPermission/fullHouseholderInConservationArea.json
