+++ examples/application/planningPermission/major.json
