+++ examples/application/planningPermission/minor.json
