+++ types/schemas/application/data/Property.ts
