+++ types/schemas/application/data/Proposal.ts
