+++ types/schemas/application/data/shared.ts
