+++ types/schemas/application/enums/ApplicationTypes.ts
