+++ types/schemas/application/enums/BuildingRegulations.ts
