+++ types/schemas/application/enums/FileTypes.ts
