+++ types/schemas/application/enums/Flags.ts
