+++ types/schemas/application/enums/OpenSpaces.ts
