+++ types/schemas/application/enums/PlanningConstraints.ts
