+++ types/schemas/application/enums/ProjectTypes.ts
