+++ types/schemas/application/enums/PropertyTypes.ts
