+++ types/schemas/application/enums/ProtectedSpaces.ts
