+++ types/schemas/application/enums/ResidentialUnitTypes.ts
