+++ types/schemas/application/enums/TenureTypes.ts
