+++ types/schemas/application/enums/UseClasses.ts
