+++ types/schemas/application/index.ts
@@ -3,19 +3,19 @@ import {Metadata} from './Metadata';
 import {PreAssessment} from './PreAssessment';
 import {Responses} from './Responses';
 import {Applicant} from './data/Applicant';
-import {Application} from './data/Application';
+import {ApplicationData} from './data/ApplicationData';
 import {FilesAsData} from './data/Files';
 import {Property} from './data/Property';
 import {Proposal} from './data/Proposal';
 import {User} from './data/User';
 
 /**
- * @title Digital Planning Application
+ * @title Application
  * @description The root specification for a planning application in England generated by a digital planning service
  */
-export interface DigitalPlanningApplication {
+export interface Application {
   data: {
-    application: Application;
+    application: ApplicationData;
     user: User;
     applicant: Applicant;
     property: Property;
